Differentiate producer/consumer configuration (#16)
The producer/consumer code was sharing the same configuration, which
means the producer was being initialized as a consumer as well. If the
producer starts first, it joins the same group and is assigned the only
partition for the topic. When the consumer eventually comes online, it
joins the group but is not assigned the partition, appearing to be
stuck. This patch bifurcates the configuration on whether or not the
Kafka client being stood up should be a consumer. Also, the PR adds a
few hygienic options that should always be set for consumers (read
committed, require stable offsets).